requirements.txt
# requirements.txt
pandas>=1.5.0
pymysql>=1.0.0
requests>=2.28.0
numpy>=1.23.0

# Yahoo Finance & Data
yfinance>=0.2.0
akshare>=1.10.0

# Database
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0

# Charting
mplfinance>=0.12.0
matplotlib>=3.7.0
